refactor: align shuffle test/perf file names with binding
bindings/test/swapOrNotShuffle.test.ts merges into shuffle.test.ts and
bindings/perf/swapOrNotShuffle.test.ts becomes perf/shuffle.test.ts,
matching bindings/napi/shuffle.zig.
